src/WeatherForecastKPDL/Services/Backend/Clustering/WeatherForecast.Clustering/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
httpx==0.26.0
pandas==2.2.0
numpy==1.26.3
apscheduler==3.10.1
kafka-python==2.0.2
pydantic-settings==2.1.0
celery==5.3.0
redis[hiredis]==4.5.0
scikit-learn==1.4.0